Moved activity models from `opengever.activity.models` to `opengever.activity.model`.
Additionally removed unnecessary constructor redefinitions.
