+++ opengever/activity/center.py
@@ -1,8 +1,8 @@
 from opengever.activity.mail import PloneNotificationMailer
-from opengever.activity.models import Activity
-from opengever.activity.models import Notification
-from opengever.activity.models import Resource
-from opengever.activity.models import Watcher
+from opengever.activity.model import Activity
+from opengever.activity.model import Notification
+from opengever.activity.model import Resource
+from opengever.activity.model import Watcher
 from opengever.base.model import create_session
 from opengever.base.oguid import Oguid
 from opengever.ogds.models.query import extend_query_with_textfilter
