+++ opengever/activity/model/__init__.py
@@ -1,0 +1,4 @@
+from opengever.activity.model.activity import Activity
+from opengever.activity.model.notification import Notification
+from opengever.activity.model.resource import Resource
+from opengever.activity.model.watcher import Watcher
