+++ opengever/testing/builders/activity.py
@@ -1,8 +1,8 @@
 from ftw.builder import builder_registry
-from opengever.activity.models import Activity
-from opengever.activity.models import Notification
-from opengever.activity.models import Resource
-from opengever.activity.models import Watcher
+from opengever.activity.model import Activity
+from opengever.activity.model import Notification
+from opengever.activity.model import Resource
+from opengever.activity.model import Watcher
 from opengever.base.oguid import Oguid
 from opengever.ogds.models.tests.builders import SqlObjectBuilder
 from plone.app.testing import TEST_USER_ID
